+++ NEWS.md
@@ -1,4 +1,9 @@
 
+
+# rmonocypher 0.1.6  2024-05-07
+
+* `encrypt_obj()` to serialize encrypted objects to file
+
 # rmonocypher 0.1.5  2024-05-07
 
 * Add `blake2b()` for hashing any R object